requirements.txt
fastapi
uvicorn
pandas
sendgrid
python-dotenv
requests
sqlalchemy
passlib[bcrypt]
python-jose
python-multipart
